docs(zh-hans/transitionprocess): update wording for cosmetic transition description
Updated the description of how cosmetic and surgical procedures can help align
physical appearance with gender identity. Changed "与必要的整形手术" to "甚至是整形手术"
to better reflect that surgery is not always necessary but can be part of the process.

docs(zh-hant/transitionprocess): update wording for cosmetic transition description

Updated the description of how cosmetic and surgical procedures can help align
physical appearance with gender identity. Changed "與必要的整形手術" to "甚至是整形手術"
to better reflect that surgery is not always necessary but can be part of the process.